Added support to (re)initialize project

+++ initProject.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-if [ -z "$1" ]
+if [ -z "$1" ];
   then
     echo "Please provide a name for your service, e.g. ./initProject.sh offer-service"
     exit
@@ -8,8 +8,12 @@ fi
 
 serviceName=$1
 
-echo "### Installing symfony skeleton"
-docker run --rm -v "$(pwd)":/app composer create-project symfony/skeleton "${serviceName}"
+if [ ! -d "${serviceName}" ];
+  then
+    echo "### Installing symfony skeleton"
+    docker run --rm -v "$(pwd)":/app composer create-project symfony/skeleton "${serviceName}"
+fi
+
 
 echo "### Installing api-platform"
 docker run --rm -v "$(pwd)"/"${serviceName}":/app composer composer require api-platform/api-pack
